chore: refactor hosting into App.xaml.host.cs
