+++ Chefs/Chefs.csproj
@@ -80,6 +80,12 @@
 		<DefineConstants Condition="'$(UseSkiaRendering)'=='true'">$(DefineConstants);HAS_SKIA_RENDERER</DefineConstants>
 	</PropertyGroup>
 
+	<ItemGroup>
+		<Compile Include="App.xaml.host.cs">
+			<DependentUpon>App.xaml</DependentUpon>
+		</Compile>
+	</ItemGroup>
+	
 	<Choose>
 		<When Condition="$(IsWinAppSdk)">
 			<ItemGroup>
